We cant delete an agency if agents are linked to it, so changes made in delete method: agents are moved to wait list. we could also delete the agents if wanted

+++ HemDotNetWebApi/Data/RealEstateAgencyRepository.cs
@@ -23,15 +23,32 @@ namespace HemDotNetWebApi.Data
             return await _context.RealEstateAgencies.Include(a => a.RealEstateAgencyAgents).ToListAsync();
         }
 
+        // Allan
         public async Task<bool> DeleteAgency(int id)
         {
-            var agency = await _context.RealEstateAgencies.FindAsync(id);
+            // We want to delete an agency, but the agency has agents; what to do?
+            // here, we decide to assign those agents to the wait list.
+            var agency = await _context.RealEstateAgencies
+                .Include(a => a.RealEstateAgencyAgents)
+                .FirstOrDefaultAsync(a => a.RealEstateAgencyId == id);
+
             if (agency == null)
-            {
                 return false;
+
+            var waitListAgency = await _context.RealEstateAgencies
+                .FirstOrDefaultAsync(a => a.RealEstateAgencyName == "Wait List");
+
+            if (waitListAgency == null)
+                throw new InvalidOperationException("Wait List agency not found. Please create one before deleting agencies.");
+
+            
+            foreach (var agent in agency.RealEstateAgencyAgents)
+            {
+                agent.RealEstateAgentAgencyId = waitListAgency.RealEstateAgencyId;
             }
 
             _context.RealEstateAgencies.Remove(agency);
+
             await _context.SaveChangesAsync();
             return true;
         }
